examples/proxy_workflows/requirements.txt
openassetio-manager-bal==1.0.0a4
